+++ HISTORY.rst
@@ -62,3 +62,4 @@
 
 * Correct random row choice for MSSQL (Joe Francis)
 * Adjust sequences only for specified schemas (R Vrajmohan)
+* Fix for arrays of ENUM (Federico Bertola)